price chart refactoring && 24h to 48h
