+++ packages/ui/public/locales/en/home.json
@@ -30,5 +30,5 @@
   "lastVotingPowerChange": "Last Voting Power Change",
   "powerChange": "<0>{{change}}</0> at block ",
   "validators": "Validators",
-  "priceHistory": "Price (~24h)"
+  "priceHistory": "Price (~48h)"
 }
