+++ packages/ui/src/screens/home/components/hero/components/token_price/styles.ts
@@ -4,9 +4,6 @@ const useStyles = makeStyles()(() => ({
   chart: {
     height: '285px',
     width: '100%',
-    '& .yAxis .recharts-cartesian-axis-tick:first-of-type': {
-      display: 'none',
-    },
   },
 }));
